Update pcgMexFile.c
To fix the issue that OPTIMIZATIONLEVEL=1 causes error. The assignment of the pointer x to the output pointer plhs[0] (plhs[0]=x;) seems illegal. Maybe in newer versions of MATLAB some memory is deallocated automatically so the above assignment no longer works.
Instead of directly assigning pointer x to plhs[0], we create a space in C, copy the data in x to the space, then let plhs[0] point to the space.